Feat/Refactor : Landing page separated from Index.tsx file and added custom loading icon instead of Loading.. text
